requirements.txt
# Core ML and Data Science
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.0.0
ta==0.10.2
ta-lib==0.4.28

# API Clients
python-binance==1.0.19
requests==2.31.0
tweepy==4.14.0
snscrape==0.7.0.20230622

# Data Processing
yfinance==0.2.18
ccxt==4.0.77

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Environment and Config
python-dotenv==1.0.0
configparser==6.0.0

# Async Support
aiohttp==3.8.5
asyncio==3.4.3

# Utilities
schedule==1.2.0
loguru==0.7.0
